+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Versioning follows a `MAJOR`.`MINOR`.`PATCH` pattern, where:
 
 - Use links and buttons consistently [#258](https://github.com/spraakbanken/mink-frontend/issues/258)
 - Remove table row striping
+- Replace highlight.js with CodeMirror
 
 ## [2.1.1](https://github.com/spraakbanken/mink-frontend/releases/tag/v2.1.1) (2026-06-12)